mark/space parity + custom baudrate
Added mark and space parity using the CMSPAR flag in combination with the PARODD flag. Added custom baudrate using the termios2 structure instead of the termios header. Added function to change the parity of a port given the device File and the parity. Added constants to set parity easier (example.. Parity_Even instead of 2).